Move TrainerPartyTemplate to its own file
