+++ src/data/mystery-encounters/encounters/weird-dream-encounter.ts
@@ -41,7 +41,8 @@ import { TrainerType } from "#enums/trainer-type";
 import PokemonData from "#app/system/pokemon-data";
 import { Nature } from "#enums/nature";
 import type HeldModifierConfig from "#app/interfaces/held-modifier-config";
-import { trainerConfigs, TrainerPartyTemplate } from "#app/data/trainers/trainer-config";
+import { trainerConfigs } from "#app/data/trainers/trainer-config";
+import { TrainerPartyTemplate } from "#app/data/trainers/TrainerPartyTemplate";
 import { PartyMemberStrength } from "#enums/party-member-strength";
 
 /** i18n namespace for encounter */
